Updates V1 to optimized pretty-printer
