+++ partiql-ast/src/main/kotlin/org/partiql/ast/normalize/Normalize.kt
@@ -1,3 +1,4 @@
+@file:JvmName("Normalize")
 /*
  * Copyright 2022 Amazon.com, Inc. or its affiliates.  All rights reserved.
  *
