+++ partiql-ast/src/main/kotlin/org/partiql/ast/sql/SqlLayout.kt
@@ -1,97 +1,78 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates.  All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License").
+ *  You may not use this file except in compliance with the License.
+ * A copy of the License is located at:
+ *
+ *      http://aws.amazon.com/apache2.0/
+ *
+ *  or in the "license" file accompanying this file. This file is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific
+ *  language governing permissions and limitations under the License.
+ */
+
 package org.partiql.ast.sql
 
 /**
  * [SqlLayout] determines how an [SqlBlock] tree is transformed in SQL text.
  */
-@Deprecated("This will be changed in the next major version")
-public abstract class SqlLayout {
+public interface SqlLayout {
 
-    public abstract val indent: Indent
-
-    public open fun format(root: SqlBlock): String {
-        val ctx = Ctx.empty()
-        root.accept(Formatter(), ctx)
-        return ctx.toString()
-    }
+    public fun print(block: SqlBlock): String
 
     public companion object {
 
         /**
-         * Default SQL format.
+         * Default SQL layout.
          */
-        public val DEFAULT: SqlLayout = object : SqlLayout() {
+        @JvmStatic
+        public val STANDARD: SqlLayout = object : SqlLayout {
 
-            override val indent = Indent(2, Indent.Type.SPACE)
+            override fun print(block: SqlBlock): String {
+                val sb = StringBuilder()
+                var curr: SqlBlock? = block
+                while (curr != null) {
+                    when (curr) {
+                        is SqlBlock.None -> {}
+                        is SqlBlock.Line -> sb.appendLine()
+                        is SqlBlock.Text -> sb.append(curr.text)
+                        is SqlBlock.Nest -> {
+                            if (curr.prefix != null) sb.append(curr.prefix)
+                            sb.append(print(curr.child))
+                            if (curr.postfix != null) sb.append(curr.postfix)
+                        }
+                    }
+                    curr = curr.next
+                }
+                return sb.toString()
+            }
         }
 
         /**
          * Write SQL statement on one line.
          */
-        public val ONELINE: SqlLayout = object : SqlLayout() {
+        @JvmStatic
+        public val ONELINE: SqlLayout = object : SqlLayout {
 
-            override val indent = Indent(2, Indent.Type.SPACE)
-
-            override fun format(root: SqlBlock): String = root.toString().replace("\n", "")
+            override fun print(block: SqlBlock): String {
+                val sb = StringBuilder()
+                var curr: SqlBlock? = block
+                while (curr != null) {
+                    when (curr) {
+                        is SqlBlock.None -> {}
+                        is SqlBlock.Line -> {} // ignore
+                        is SqlBlock.Text -> sb.append(curr.text)
+                        is SqlBlock.Nest -> {
+                            if (curr.prefix != null) sb.append(curr.prefix)
+                            sb.append(print(curr.child))
+                            if (curr.postfix != null) sb.append(curr.postfix)
+                        }
+                    }
+                    curr = curr.next
+                }
+                return sb.toString()
+            }
         }
-    }
-
-    /**
-     * [SqlLayout] indent configuration.
-     *
-     * @property count
-     * @property type
-     */
-    public class Indent(
-        public val count: Int,
-        public val type: Type,
-    ) {
-
-        public enum class Type(public val char: Char) {
-            TAB(Char(9)),
-            SPACE(Char(32)), ;
-        }
-
-        override fun toString(): String = type.char.toString().repeat(count)
-    }
-
-    private class Ctx private constructor(val out: StringBuilder, val level: Int) {
-        fun nest() = Ctx(out, level + 1)
-
-        override fun toString() = out.toString()
-
-        companion object {
-            fun empty() = Ctx(StringBuilder(), 0)
-        }
-    }
-
-    private inner class Formatter : BlockBaseVisitor<Unit, Ctx>() {
-
-        private inline fun write(ctx: Ctx, f: () -> String) {
-            if (ctx.level > 0) ctx.out.append(lead(ctx))
-            ctx.out.append(f())
-        }
-
-        override fun defaultReturn(block: SqlBlock, ctx: Ctx) = write(ctx) {
-            block.toString()
-        }
-
-        override fun visitNil(block: SqlBlock.Nil, ctx: Ctx) {}
-
-        override fun visitNewline(block: SqlBlock.NL, ctx: Ctx) {
-            ctx.out.appendLine()
-        }
-
-        override fun visitText(block: SqlBlock.Text, ctx: Ctx) = write(ctx) { block.text }
-
-        override fun visitNest(block: SqlBlock.Nest, ctx: Ctx) {
-            block.child.accept(this, ctx.nest())
-        }
-
-        override fun visitLink(block: SqlBlock.Link, ctx: Ctx) {
-            block.lhs.accept(this, ctx)
-            block.rhs.accept(this, ctx)
-        }
-
-        private fun lead(ctx: Ctx) = indent.toString().repeat(ctx.level)
     }
 }
